Move async batch iterator to own file set
